[WIP] extracted BackendNodeHandles.hpp
